### 2018-05-20 #743 @whedon
[PRE REVIEW]: sjmisc - Data and Variable Transformation Functions
**Submitting author:** @strengejacke (<a href="http://orcid.org/0000-0002-8895-3206">Daniel Lüdecke</a>)
**Repository:** <a href="https://github.com/strengejacke/sjmisc" target ="_blank">https://github.com/strengejacke/sjmisc</a>
**Version:** 2.7.2
