### 2018-05-20 #743 @arfon
@strengejacke - this PR should fix your paper metadata https://github.com/strengejacke/sjmisc/pull/85
