### 2018-05-20 #743 @jrosen48
I'd be pleased to review **sjmisc**.
