### 2018-05-20 #743 @leeper
@methodds Can you confirm you're interested in reviewing this for JOSS?
