### 2018-05-20 #743 @leeper
@strengejacke Can you merge [this PR](https://github.com/strengejacke/sjmisc/pull/87) to fix your bibliography?
