### 2018-05-27 #743 @leeper
@jrosen48 and @methodds thanks for agreeing to review! You each have a review checklist to complete over at https://github.com/openjournals/joss-reviews/issues/754

Let me know if you have any questions.
