### 2018-05-28 #754 @jrosen48
@leeper if I have comments about, just for example because it's the first item I came upon that I think I may have a note for, the following item:

`Repository: Is the source code for this software available at the repository url?

Do you suggest I add them a) in a comment here below the checklist, b) as a note beside the checklist item, c) as an issue in the repository, d) a differeny way?

Thanks!
