### 2018-05-28 #754 @leeper
@jrosen48 You can do either (a) open an issue on the source repo with a link back to this page, asking for changes, or (b) create a cumulative review that raises all issues not addressed in the checklist and leave it as a comment here, or a mix of both.
