### 2018-05-29 bbeef2e
fix issue in count_na()
