+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: sjmisc
 Type: Package
 Encoding: UTF-8
 Title: Data and Variable Transformation Functions
-Version: 2.7.2
+Version: 2.7.2.9000
 Date: 2018-05-04
 Author: Daniel Lüdecke <d.luedecke@uke.de>
 Authors@R: person("Daniel", "Lüdecke", role = c("aut", "cre"), email = "d.luedecke@uke.de", comment = c(ORCID = "0000-0002-8895-3206"))
