### 2018-05-31 d49071c
new function is_cross_classified()
