+++ NAMESPACE
@@ -70,6 +70,7 @@ export(group_labels_if)
 export(group_str)
 export(group_var)
 export(group_var_if)
+export(is_cross_classified)
 export(is_crossed)
 export(is_empty)
 export(is_even)
