### 2018-06-01 0426308
fix issue in merge_imputation()
