### 2018-06-05 cbe463b
improve ref_lvl()
