### 2018-06-06 f9eadc1
remove redundant DESCRIPTION field
