### 2018-06-11 #754 @strengejacke
May I kindly ask how the review process continues? Will the reviewers let me or you know when they finished their review, or will I be contacted via notification? Just wondering, because from the checklist you can't completely decide if there are parts left to review, or if these are issues I have to address. :-)
