### 2018-06-11 strengejacke/sjmisc#90 @leeper
JOSS: Add CONTRIBUTING file
Please add a CONTRIBUTING.md file, or similar, describing how users can/should contribute to the package.

For JOSS: https://github.com/openjournals/joss-reviews/issues/754
