### 2018-06-11 strengejacke/sjmisc#91 @leeper
JOSS: Add details to README about need
The paper has a good, brief summary of need. Please add a few sentences to the README that provides a similar motivation for the package.

For JOSS: https://github.com/openjournals/joss-reviews/issues/754
