### 2018-06-11 strengejacke/sjmisc#92 @leeper
JOSS: Add tests
While the package has thorough examples, please add a test suite in `/tests` that provides coverage of most of the functionality in the package. At a minimum it should provides tests demonstrating that functions run with the documented arguments; even better would be thorough tests that check whether function outputs are correct and match documented return values.

For JOSS: https://github.com/openjournals/joss-reviews/issues/754
