### 2018-06-11 #754 @leeper
@strengejacke  I've had a look at the reviews so far and opened several issues for you to address in line with the unchecked items above. Please address these and I will then ask the reviewers to take another look.
