+++ README.md
@@ -2,7 +2,9 @@
 
 [![DOI](https://zenodo.org/badge/32622582.svg)](https://zenodo.org/badge/latestdoi/32622582)
 
-Collection of miscellaneous utility functions, supporting data transformation tasks like recoding, dichotomizing or grouping variables, setting and replacing missing values. The data transformation functions also support labelled data.
+Data preparation is a common task in research, which usually takes the most amount of time in the analytical process. Packages for data preparation have been released recently as part of the _tidyverse_, focussing on the transformation of data sets. Packages with special focus on transformation of _variables_, which fit into the workflow and design-philosophy of the tidyverse, are missing.
+
+``sjmisc`` tries to fill this gap. Basically, this package complements the ``dplyr`` package in that ``sjmisc`` takes over data transformation tasks on variables, like recoding, dichotomizing or grouping variables, setting and replacing missing values, etc. A distinctive feature of ``sjmisc`` is the support for labelled data, which is especially useful for users who often work with data sets from othert statistical software packages like _SPSS_ or _Stata_.
 
 The functions of **sjmisc** are designed to work together seamlessly with other packes from the tidyverse, like **dplyr**. For instance, you can use the functions from **sjmisc** both within a pipe-workflow to manipulate data frames, or to create new variables with `mutate()`. See `vignette("design_philosophy", "sjmisc")` for more details.
 
