### 2018-06-15 ea22307
typo in docs
