### 2018-06-15 acf8e77
add tests
