### 2018-06-15 e984fca
add tests
