### 2018-06-16 aba8137
fix issue in find_var(), add tests
