### 2018-06-16 8bad16f
add tests
