### 2018-06-16 a33e18c
add tests
