### 2018-06-16 #754 @strengejacke
Thanks for your comments, I have addressed the above mentioned issues.

1) I added a paragraph with statement of need to the README file, so it's immediately obvious to visitors of the GitHub Repo. Furthermore, the README now also links to a CONTRIBUTION file.

2) I added a detailed CONTRIBUTION file, which guides users how they may contribute to my package.

3) I added a lot of tests (using testthat) to the `/tests`-folder. Tests not just check if arguments are working (which is also done in the examples), but check if certain inputs result in the expected outputs.
