### 2018-06-17 #754 @leeper
Thanks. @jrosen48 can you take another look and see if anything else should be addressed?
