### 2018-06-18 4fab59d
fix title in paper.md
