### 2018-06-20 #754 @jrosen48
Hi @leeper, I think @strengejacke did a great job of addressing the issues you raised.

1. I think the paragraph with a statement of need added to the README is strong (particularly this: "Packages for data preparation have been released recently as part of the tidyverse, focussing on the transformation of data sets. Packages with special focus on transformation of variables, which fit into the workflow and design-philosophy of the tidyverse, are missing.").

2. The contribution page is excellent, with a description of *multiple* ways users can contribute, including filing issues and making pull requests (with specific and helpful instructions in both cases).

3. The tests use built-in and small, constructed data sets, to test many (most?) to test multiple sets of arguments and types of inputs to, as @strengejacke mentioned, test their output.
