### 2018-06-20 #754 @strengejacke
At this point I'd really like to thank the editors and reviewers for their work so far. The process of paper submission and review raised awareness of the fact that a GitHub-repo is not just only a "cloud-backup", but also a place where users inform themselves about software and its intention and need. I think changing the perspective from the repo-owner to a "visitor" - and that's what's necessary to do when preparing a paper for JOSS - is really helpful to make the software and documentation more accessable!
