### 2018-06-20 #754 @leeper
@jrosen48 Thank you! It sounds like we're both satisfied then.

@strengejacke Please issue a release of the package and obtain a DOI for the release from, for example, figshare or zenodo and post that DOI here when you have it.
