### 2018-06-20 #754 @strengejacke
The current version (2.7.2) is already registered at zenodo: http://doi.org/10.5281/zenodo.1249192
The doi referring to the latest version on zenodo is 10.5281/zenodo.1249191.
