### 2018-06-20 #754 @leeper
You should be able to get a DOI for a new release. We want the one for the final version as accepted by JOSS.
