### 2018-06-20 #754 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
