### 2018-06-20 #754 @strengejacke
Thanks a lot to all! btw, I signed up as potential reviewer already some time ago, I really like the idea of this journal. Especially thanks to @leeper - my impression is, that he's the only editor for social sciences and R and has to edit quite some submissions here. Great work!
