### 2018-06-20 #754 @jrosen48
Congrats @strengejacke--please keep up the exemplary work grealt benefits the R community. Thanks  @leeper and @methodds & @arfon for being part of a great process (for reviewers!).
